refactor: CartPage 퍼널 구조 제거
